Rename ThemeProvider -> PenumbraUIProvider
