+++ packages/ui/src/Button/index.test.tsx
@@ -1,14 +1,14 @@
 import { describe, expect, it, vi } from 'vitest';
 import { Button } from '.';
 import { fireEvent, render } from '@testing-library/react';
-import { ThemeProvider } from '../ThemeProvider';
+import { PenumbraUIProvider } from '../PenumbraUIProvider';
 import { Check } from 'lucide-react';
 
 describe('<Button />', () => {
   it('calls the passed-in click handler when clicked', () => {
     const onClick = vi.fn();
     const { getByText } = render(<Button onClick={onClick}>Click me</Button>, {
-      wrapper: ThemeProvider,
+      wrapper: PenumbraUIProvider,
     });
 
     fireEvent.click(getByText('Click me'));
@@ -18,7 +18,7 @@ describe('<Button />', () => {
 
   describe('when `iconOnly` is falsey', () => {
     it('renders `children` as the button text', () => {
-      const { queryByText } = render(<Button>Label</Button>, { wrapper: ThemeProvider });
+      const { queryByText } = render(<Button>Label</Button>, { wrapper: PenumbraUIProvider });
 
       expect(queryByText('Label')).toBeTruthy();
     });
@@ -30,7 +30,7 @@ describe('<Button />', () => {
         <Button iconOnly icon={Check}>
           Label
         </Button>,
-        { wrapper: ThemeProvider },
+        { wrapper: PenumbraUIProvider },
       );
 
       expect(queryByText('Label')).toBeNull();
@@ -42,7 +42,7 @@ describe('<Button />', () => {
         <Button iconOnly icon={Check}>
           Label
         </Button>,
-        { wrapper: ThemeProvider },
+        { wrapper: PenumbraUIProvider },
       );
 
       expect(queryByTitle('Label')).toBeTruthy();
