+++ packages/ui/src/PenumbraUIProvider/fonts/WorkSans-Medium-Vietnamese.woff2
