+++ packages/ui/src/Pill/index.test.tsx
@@ -1,11 +1,11 @@
 import { describe, expect, it } from 'vitest';
 import { Pill } from '.';
 import { render } from '@testing-library/react';
-import { ThemeProvider } from '../ThemeProvider';
+import { PenumbraUIProvider } from '../PenumbraUIProvider';
 
 describe('<Pill />', () => {
   it('renders its `children`', () => {
-    const { queryByText } = render(<Pill>Contents</Pill>, { wrapper: ThemeProvider });
+    const { queryByText } = render(<Pill>Contents</Pill>, { wrapper: PenumbraUIProvider });
 
     expect(queryByText('Contents')).toBeTruthy();
   });
